Rename folders and files names in homepage
