--- frontend/src/helpers.js
@@ -1,3 +1,0 @@
-export const createExtension = (file)=>{
-    return file.name.substr(file.name.lastIndexOf('.'));
-}
