+++ frontend/src/views/HomePage/components/ImagesSection/ImagesSection.styles.js
